Refactor populatePresetSelector controller func
Separate data retrieval, event-handler attachment, and view-updating components
